requirements.txt
django-ckeditor==5.7.1 # Version 5.2.2 to support PasteFromWord into CKEditor
XlsxWriter==0.9.3
voluptuous==0.9.3
weasyprint==0.42.3
attrs==19.3.0

coverage==6.1.1
pycodestyle==2.4.0
pylint==2.4.4
diff-cover==6.4.2
tblib==1.7.0